redesigned the code
