+++ .env.template
@@ -82,8 +82,8 @@ AIRBYTE_SERVER_APIVER=
 AIRBYTE_API_TOKEN=
 AIRBYTE_DESTINATION_TYPES=
 AIRBYTE_OAUTH_REDIRECT_URL=
-AIRBYTE_GSHEETS_OAUTH_CLIENT_ID=
-AIRBYTE_GSHEETS_OAUTH_CLIENT_SECRET=
+AIRBYTE_GOOGLE_OAUTH_CLIENT_ID=
+AIRBYTE_GOOGLE_OAUTH_CLIENT_SECRET=
 AIRBYTE_FETCH_CONNECTION_CATALOG_TIMEOUT_SECONDS=300
 AIRBYTE_SOURCE_BLACKLIST=""
